Add ability to test payments with Tiamat
This commit updates the testing data model to include verfication of payments
and adds asynchronous task waiting for payment notification messages.
It also enforces the usage of certificate pinning in order
to talk to a full-service api account for the testing account's
wallet. It also allows profile setting via Auxin.